Separate Homepage into components
Implementation Details

1. Each component properly implements i18n support
2. State management uses Vue 3 reactivity system with refs
3. Components follow accessibility guidelines (proper ARIA attributes)
4. Tailwind classes maintain consistency across components
5. Proper TypeScript typing for props and emits

The code now follows modern Vue 3.4+ idioms while remaining functionally equivalent to the original implementation.
