+++ src/pages/index.astro
@@ -1,6 +1,6 @@
 ---
 import Layout from "../layouts/Layout.astro";
-import Homepage from "../components/attempt3/Homepage.vue";
+import Homepage from "../components/attempt1/Homepage.vue";
 import "../styles/global.css";
 import "../styles/tailwind.css";
 import "../fonts/font-face.css";
